Move the result component to entities/result to improve performance (not having to import the whole courses module to reuse the component)
