+++ src/main/webapp/app/assessment-dashboard/assessment-dashboard.component.ts
@@ -8,7 +8,7 @@ import { ExerciseResultService } from '../entities/result/result.service';
 import { DifferencePipe } from 'angular2-moment';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Result } from '../entities/result';
-import { ResultDetailComponent } from '../courses/results/result.component';
+import { ResultDetailComponent } from '../entities/result/result.component';
 import { ModelingAssessmentService } from '../entities/modeling-assessment/modeling-assessment.service';
 import { HttpResponse } from '@angular/common/http';
 import { Principal } from '../shared';
