+++ src/main/webapp/app/assessment-dashboard/assessment-dashboard.module.ts
@@ -7,9 +7,7 @@ import { JhiMainComponent } from '../layouts';
 import { AssessmentDashboardComponent } from './assessment-dashboard.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { SortByModule } from '../components/pipes/sort-by.module';
-import { ResultDetailComponent } from '../courses';
-import { ResultComponent } from '../courses/results/result.component';
-import { ArTEMiSCoursesModule } from '../courses/courses.module';
+import { ArTEMiSResultModule, ResultComponent, ResultDetailComponent } from '../entities/result';
 
 const ENTITY_STATES = [
     {
@@ -29,7 +27,7 @@ const ENTITY_STATES = [
         RouterModule.forChild(ENTITY_STATES),
         NgbModule,
         SortByModule,
-        ArTEMiSCoursesModule
+        ArTEMiSResultModule
     ],
     declarations: [
         AssessmentDashboardComponent,
