+++ src/main/webapp/app/courses/index.ts
@@ -2,4 +2,3 @@ export * from './courses.component';
 export * from './courses.module';
 export * from './courses.route';
 export * from './exercises/exercise-list.component';
-export * from './results/result.component';
