+++ src/main/webapp/app/entities/result/index.ts
@@ -1,2 +1,4 @@
 export * from './result.model';
 export * from './result.service';
+export * from './result.component';
+export * from './result.module';
