+++ src/main/webapp/app/entities/result/result-detail.html
