+++ src/main/webapp/app/entities/result/result.component.html
