+++ src/main/webapp/app/entities/result/result.module.ts
@@ -1,14 +1,26 @@
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 
 import { ArTEMiSSharedModule } from '../../shared';
-import {
-    ResultService
-} from './';
+import { ResultService } from './';
 import { ExerciseResultService } from './result.service';
+import { ResultComponent, ResultDetailComponent } from './result.component';
+import { MomentModule } from 'angular2-moment';
 
 @NgModule({
     imports: [
-        ArTEMiSSharedModule
+        ArTEMiSSharedModule,
+        MomentModule
+    ],
+    declarations: [
+        ResultComponent,
+        ResultDetailComponent
+    ],
+    exports: [
+        ResultComponent
+    ],
+    entryComponents: [
+        ResultComponent,
+        ResultDetailComponent
     ],
     providers: [
         ResultService,
